src/requirements.txt
flask==3.0.3
waitress==3.0.0
python-dotenv==1.0.1